packages/backend/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pytz==2021.3
sqlparse==0.4.2
django-cors-headers==3.11.0
# postgres
psycopg2-binary==2.9.3

